Got a new dialect driver and have been fixing things

+++ src/main/java/net/draconia/jobsemailcollector/dao/IndividualDAO.java
@@ -11,7 +11,7 @@ public interface IndividualDAO extends Serializable
 	public void delete(final Individual objIndividual);
 	public Individual getIndividualById(final Integer iId);
 	public Individual getIndividualByName(final String sName);
-	public Integer getIndividualCount();
+	public Long getIndividualCount();
 	public List<Individual> list();
 	public void save(final Individual objIndividual);
 }
