+++ src/main/java/net/draconia/jobsemailcollector/service/IndividualService.java
@@ -10,7 +10,7 @@ public interface IndividualService extends Serializable
 	public void delete(final Individual objIndividual);
 	public Individual getIndividualById(final Integer iId);
 	public Individual getIndividualByName(final String sName);
-	public Integer getIndividualCount();
+	public Long getIndividualCount();
 	public List<Individual> list();
 	public void save(final Individual objIndividual);
 }
